Collection processors take IEnumerable; sum() folds with +; nullable custom operators compare
ICollectionProcessor.Process now takes IEnumerable rather than the non-generic
ICollection. A HashSet<T>, a declared ISet<T> and a bare IEnumerable<T> do not
implement that interface, so every processor over one threw for the interpreter
while the compiled path answered, and which a caller got followed from their
declared context type. This is a breaking change for anyone implementing their
own processor: one signature, and a count without walking is available from
CollectionOperandUtils.TryGetCount. count() also gets its O(1) answer back for
every source that can give one, by asking the runtime type about both counting
interfaces instead of the declared type about one.

sum() is a fold of +, so the accumulator starts as the first item and the
operator's own promotion decides the running type. An int collection answers
Int32 rather than Double, an empty one answers a zero of its item type, and the
compiled integral sums are unchecked to match +. average() is untouched, since a
quotient has to accumulate in a real type. One expression stops working: summing
a collection holding both an int and a ulong throws now, because 1 + someUlong
has always been refused.

A nullable operand reaches its own type's relational operator. Ordered? < Ordered
compiled and then threw at evaluation; the lookup is asked again on the operands
the nullable machinery unwraps. An empty nullable still sorts before everything.
between deliberately does not consult these operators, because the interpreter
cannot.

+++ SpringExpressionsTests/Expressions/ProjectionSelectionAgreementTests.cs
@@ -424,10 +424,10 @@ namespace SpringExpressionsTests.Expressions
         /// type, which is what lets the aggregation stay compiled.
         /// </summary>
         /// <remarks>
-        /// Asserted per backend, because the backends disagree on sum()'s result type for ints - int
-        /// compiled, double interpreted. That divergence belongs to the numeric-promotion cluster, not
-        /// to projections; this test pins only that the projection under the aggregator keeps working
-        /// from both sides.
+        /// This used to be asserted per backend, because <c>sum()</c> over ints answered <c>Int32</c>
+        /// compiled and <c>Double</c> interpreted - the interpreter's accumulator seeded at <c>0d</c>
+        /// whatever the item type. That is ruled now (<i>sum() is a fold of '+'</i>), so both answer
+        /// <c>Int32</c> and the test asserts agreement like every other one here.
         /// </remarks>
         [Test]
         public void AggregatorOverAProjectionStaysCompiledAtTheItemType()
@@ -439,8 +439,8 @@ namespace SpringExpressionsTests.Expressions
             Assert.AreEqual(9, compiled);
 
             var interpreted = InterpretGetter<ProjectionSourceHolder, object>("Ints.!{#this + 1}.sum()").GetValue(holder);
-            Assert.AreEqual(typeof(double), interpreted.GetType());
-            Assert.AreEqual(9.0d, interpreted);
+            Assert.AreEqual(typeof(int), interpreted.GetType());
+            Assert.AreEqual(9, interpreted);
         }
     }
 }
